# [C++] Optimize Varint paring for c++